Module holochain_net_connection::json_protocol

This module provides a higher level interface to p2p / network messaging basically handles serialization / deserialization from / to the core protocol message types (NamedBinary and Json).

Structs

ConfigData

ConnectData

DhtMetaData
Generic DHT metadata message

DropEntryData
Data Request from own p2p-module

DropMetaData
Data Request from own p2p-module

EntryData
Generic DHT data message

EntryListData

FailureResultData

FetchEntryData
Data Request from some other agent

FetchEntryResultData
DHT data response from a request

FetchMetaData
Metadata Request from another agent

FetchMetaResultData

GetListData

MessageData

MetaListData

PeerData

StateData

SuccessResultData

TrackDnaData

Enums

JsonProtocol
Enum holding all message types that serialize as json in the 'hc-core <-> P2P network module' protocol. There are 4 categories of messages: