Enum holochain_net_connection::protocol::Protocol

pub enum Protocol {
    NamedBinary(NamedBinaryData),
    Json(JsonString),
    Ping(PingData),
    Pong(PongData),
    P2pReady,
}

Low-level interface spec for communicating with the p2p abstraction notice this is not Serializable or Deserializable rmp_serde doesn't serialize enums very well... it uses indexes and arrays which are not (easily) compatible with other endpoints we use to/from NamedBinaryData to provide our own serialization wrapper

Variants
NamedBinary(NamedBinaryData)

send/recv binary data / i.e. encryption, signature messages

Json(JsonString)

send/recv generic json as utf8 strings

Ping(PingData)

send/recv a Ping message (ipc protocol spec)

Pong(PongData)

send/recv a Pong message (ipc protocol spec)

P2pReady

we have connected / configured the connection, ready for messages

Methods

impl Protocol

pub fn is_named_binary(&self) -> bool

pub fn as_named_binary<'a>(&'a self) -> &'a NamedBinaryData

pub fn is_json(&self) -> bool

pub fn as_json<'a>(&'a self) -> &'a JsonString

pub fn is_ping(&self) -> bool

pub fn as_ping<'a>(&'a self) -> &'a PingData

pub fn is_pong(&self) -> bool

pub fn as_pong<'a>(&'a self) -> &'a PongData

pub fn as_json_string(&self) -> JsonString

get a json string straight out of the Protocol enum