Enum holochain_net_connection::protocol_wrapper::ProtocolWrapper

pub enum ProtocolWrapper {
    RequestState,
    State(StateData),
    RequestDefaultConfig,
    DefaultConfig(ConfigData),
    SetConfig(ConfigData),
    Connect(ConnectData),
    PeerConnected(PeerData),
    SendMessage(MessageData),
    SendResult(MessageData),
    HandleSend(MessageData),
    HandleSendResult(MessageData),
    TrackApp(TrackAppData),
    SuccessResult(SuccessResultData),
    FailureResult(FailureResultData),
    GetDht(GetDhtData),
    GetDhtResult(DhtData),
    PublishDht(DhtData),
    StoreDht(DhtData),
    GetDhtMeta(GetDhtMetaData),
    GetDhtMetaResult(DhtMetaData),
    PublishDhtMeta(DhtMetaData),
    StoreDhtMeta(DhtMetaData),
}

High level p2p / network message

Variants
RequestState

[send] request the current state from the p2p module

State(StateData)

[recv] p2p module is telling us the current state

RequestDefaultConfig

[send] request the default config from the p2p module

DefaultConfig(ConfigData)

[recv] the default config from the p2p module

SetConfig(ConfigData)

[send] set the p2p config

Connect(ConnectData)

[send] connect to the specified multiaddr

PeerConnected(PeerData)

[recv] notification of a peer connected

SendMessage(MessageData)

[send] send a message to another node on the network

SendResult(MessageData)

[recv] recv the response back from a previous SendMessage

HandleSend(MessageData)

[recv] another node has sent us a message

HandleSendResult(MessageData)

[send] send our response to a previous HandleSend

TrackApp(TrackAppData)

[send] send out a "trackApp" request

SuccessResult(SuccessResultData)

[send / recv] report success for a messages with _id parameter

FailureResult(FailureResultData)

[send / recv] for any message with _id parameter to indicate failure

GetDht(GetDhtData)

[send] request data from the dht [recv] another node, or the network module itself is requesting data from us... send a GetDhtResult message back

GetDhtResult(DhtData)

[recv] response from requesting dht data from the network [send] success response if network is requesting this data of us

PublishDht(DhtData)

[send] publish content to the dht

StoreDht(DhtData)

[recv] the network is requesting that we store this data

GetDhtMeta(GetDhtMetaData)

[send] request meta data from the dht [recv] another node, or the network module itself is requesting data from us... send a GetDhtResult message back

GetDhtMetaResult(DhtMetaData)

[recv] response from requesting meta dht data from the network [send] success response if network is requesting this data of us

PublishDhtMeta(DhtMetaData)

[send] publish meta content to the dht

StoreDhtMeta(DhtMetaData)

[recv] the network is requesting that we store this meta data